+++ src/main/java/esgi/easisell/service/interfaces/ISaleItemService.java
@@ -1,11 +1,13 @@
 package esgi.easisell.service.interfaces;
 
 import esgi.easisell.dto.SaleItemResponseDTO;
+
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public interface ISaleItemService {
-    SaleItemResponseDTO addProductToSale(UUID saleId, String barcode, int quantity);
-    SaleItemResponseDTO addProductByIdToSale(UUID saleId, UUID productId, int quantity);
-    SaleItemResponseDTO updateItemQuantity(UUID saleItemId, int newQuantity);
+    SaleItemResponseDTO addProductToSale(UUID saleId, String barcode, BigDecimal quantity); // ✅ BigDecimal
+    SaleItemResponseDTO addProductByIdToSale(UUID saleId, UUID productId, BigDecimal quantity); // ✅ BigDecimal
+    SaleItemResponseDTO updateItemQuantity(UUID saleItemId, BigDecimal newQuantity); // ✅ BigDecimal
     void removeItemFromSale(UUID saleItemId);
 }
